+++ CHANGELOG.md
@@ -19,6 +19,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Upgraded stable Flutter analysis SDK to `2.10.3`.
  * Upgraded pana to `0.21.8`.
  * NOTE: removed `ScoreCardData.reportTypes`
+ * NOTE: started to populate `Package.withheld` and `Package.withheldExpires`.
 
 ## `20220302t144000-all`
  * Bumped runtimeVersion to `2022.02.25`.